requirements.txt
requests
websocket-client
orjson
rich
numpy
pandas
tables
h5py